nlp/requirements-lambda.txt
mangum
fastapi
unidecode
requests
python-dotenv
spacy